Finished fetchAllMessages command
The goals of this project was reduced and didn't need to copy pictures and other misc attachments. Also didn't need to combine several channels into one file so that made things simpler.

Changelog for this commit:
handled error if no messages retrieved
added counter to see progress
changed to use nickname if it exists
commented out the json code for later use
changed to write to formatted text file from object notation and upload that instead